PMM-4474 Make tests compile.
